Added a front-end dark/light mode toggler
Added a JS script that will change the theme from light to dark and vice versa on the front-end. This means that although the client-side will update the theme, user would have to reupdate the screen every time they move to a new page. This feature will be implemented shortly.